pyproject.toml
[tool.black]
line-length = 88

[tool.isort]
line_length = 88
profile = "black"
skip_glob = "venv"
known_first_party = ["raytracer", "tests"]

[tool.mypy]
exclude = "venv/"
check_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true

[tool.coverage.run]
omit = [
    "raytracer/core/config.py",
    "raytracer/cli.py",
    "raytracer/**/cli/__init__.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.poetry]
name = "raytracer"
version = "0.1.0"
description = ""
authors = ["Gary Prescott <gary@madplum.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
pillow = "^9.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.3.0"
flake8 = "^6.0.0"
flake8-isort = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.1.1"
pytest-cov = "^4.0.0"
ipython = "^8.12.0"
ipdb = "^0.13.13"
pytest-mock = "^3.10.0"
types-pillow = "^9.4.0.19"

[tool.poetry.scripts]
raytracer = "raytracer.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
